move test classes to separate files
