+++ tests/test_file_system_cache.py
@@ -1,9 +1,9 @@
 from time import sleep
 
 import pytest
-from conftest import ClearTests
-from conftest import CommonTests
-from conftest import HasTests
+from clear import ClearTests
+from common import CommonTests
+from has import HasTests
 
 from cachelib import FileSystemCache
