+++ tests/test_simple_cache.py
@@ -1,8 +1,8 @@
 from time import sleep
 
 import pytest
-from conftest import CommonTests
-from conftest import HasTests
+from common import CommonTests
+from has import HasTests
 
 from cachelib import SimpleCache
